Updated the encryption/decryption code to now support files larger than 2Gb.
