+++ Unikeys.Gui/AssemblyInfo.cs
@@ -1,5 +1,4 @@
 using System.Reflection;
-using System.Resources;
 using System.Windows;
 
 [assembly: ThemeInfo(
@@ -11,4 +10,4 @@ using System.Windows;
     // app, or any theme specific resource dictionaries)
 )]
 // The program will pull this version to display in the "About" tab
-[assembly: AssemblyVersion("1.1.2.*")]
+[assembly: AssemblyVersion("1.1.3.*")]
